arreglado detalle en vistas index

+++ database/migrations/2019_06_03_232400_create_tarjetas_table.php
@@ -17,7 +17,7 @@ class CreateTarjetasTable extends Migration
             $table->bigIncrements('id');
             $table->string('numero_tarjeta',191);
             $table->string('tipo_monedero',191);
-            $table->bigInteger('saldo');
+            $table->float('saldo');
             $table->string('benefactor',191);
             $table->timestamps();
         });
